requirements.txt
# Hyper-Agent
# Python 3.10+

pyyaml>=6.0
openai>=1.0.0
requests>=2.28.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
# Optional: Gmail read-only (enable in config)
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
